add paging

+++ src/ui.rs
@@ -13,7 +13,7 @@ pub fn render(app: &mut App, frame: &mut Frame) {
     app.table_height = layout[0].height - 1;
     let table_slice = &app
         .data_frame
-        .slice(app.table_offset.0, app.table_height as usize);
+        .slice(app.table_offset.0 as i64, app.table_height as usize);
     frame.render_stateful_widget(
         Table::new(
             rows_from_dataframe(table_slice),
